requirements.txt
Flask==3.1.2
requests==2.32.5
google-generativeai==0.3.2